I think I finally have this written so it won't hang ever:
the key is to have the receivers set a promise when they'vegotten their
last functional message and then wait for a message that tells them to
exit.

The timed part is up until the futures for all are valid.

After that, a sleep for a couple of seconds to be sure tall receivers
are ready for their exit message and then send exactly one so at which
point receivers close their socket and exit.  This ensures no messages
are sent when part of the bus has disappeard.